pyproject.toml
[project]
name = "focusbuddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-gigachat>=0.3.10",
    "pillow>=11.1.0",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
